+++ src/Powercord/plugins/pc-announcements/index.js
@@ -1,7 +1,7 @@
 const { resolve } = require('path');
 const { existsSync } = require('fs');
 const { unlink } = require('fs').promises;
-const Plugin = require('powercord/Plugin');
+const { Plugin } = require('powercord/entities');
 const { shell: { openExternal } } = require('electron');
 const { inject, uninject } = require('powercord/injector');
 const { React, ReactDOM, getModule } = require('powercord/webpack');
@@ -15,7 +15,7 @@ module.exports = class Announcements extends Plugin {
     this.notices = [];
   }
 
-  async start () {
+  async pluginDidLoad () {
     this._patchNotices();
     const injectedFile = resolve(__dirname, '..', '..', '..', '__injected.txt');
     if (existsSync(injectedFile)) {
@@ -45,7 +45,7 @@ module.exports = class Announcements extends Plugin {
     });
   }
 
-  unload () {
+  pluginWillUnload () {
     uninject('pc-custom-notices');
   }
