+++ src/Powercord/plugins/pc-announcements/manifest.json
@@ -4,7 +4,6 @@
   "description": "Broadcast announcements from Powercord API or plugins",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "dependencies": [
     "pc-classNameNormalizer"
   ]
