+++ src/Powercord/plugins/pc-badges/manifest.json
@@ -3,6 +3,5 @@
   "version": "1.0.0",
   "description": "Add some fancy custom badges",
   "author": "Powercord Team",
-  "license": "MIT",
-  "repo": null
+  "license": "MIT"
 }
