+++ src/Powercord/plugins/pc-codeblocks/manifest.json
@@ -2,7 +2,6 @@
   "name": "Better Codeblocks",
   "version": "1.0.0",
   "description": "Improve Discord codeblocks",
-  "author": "Powercord Team & noodlebox",
-  "license": "MIT",
-  "repo": null
+  "author": "Powercord Team & ThuverX#9814",
+  "license": "MIT"
 }
