+++ src/Powercord/plugins/pc-commands/manifest.json
@@ -4,7 +4,6 @@
   "description": "Powercord's command manager",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "dependencies": [
     "pc-classNameNormalizer"
   ]
