+++ src/Powercord/plugins/pc-copyRoleID/index.js
@@ -1,4 +1,4 @@
-const Plugin = require('powercord/Plugin');
+const { Plugin } = require('powercord/entities');
 const { contextMenu, getModule, getModuleByDisplayName, React } = require('powercord/webpack');
 const { waitFor, getOwnerInstance } = require('powercord/util');
 const { ContextMenu } = require('powercord/components');
@@ -6,9 +6,14 @@ const { inject, uninject } = require('powercord/injector');
 const { clipboard } = require('electron');
 
 module.exports = class CopyRoleID extends Plugin {
-  async start () {
+  async pluginDidLoad () {
     this.injectGuildRole();
     this.injectMemberRole();
+  }
+
+  pluginWillUnload () {
+    uninject('pc-guildRole');
+    uninject('pc-memberRole');
   }
 
   async injectGuildRole () {
@@ -49,10 +54,5 @@ module.exports = class CopyRoleID extends Plugin {
         })
       );
     };
-  }
-
-  unload () {
-    uninject('pc-guildRole');
-    uninject('pc-memberRole');
   }
 };
