+++ src/Powercord/plugins/pc-copyRoleID/manifest.json
@@ -3,6 +3,5 @@
   "version": "1.0.0",
   "description": "Adds a context menu to Roles in the Server Settings that allows copying the ID.",
   "author": "MrLar#2565",
-  "license": "MIT",
-  "repo": null
+  "license": "MIT"
 }
