+++ src/Powercord/plugins/pc-dnt/manifest.json
@@ -3,6 +3,5 @@
   "version": "1.0.0",
   "description": "Blocks all requests to Discord analytics endpoints",
   "author": "adam & rauenzi",
-  "license": "MIT",
-  "repo": null
+  "license": "MIT"
 }
