+++ src/Powercord/plugins/pc-guilds/manifest.json
@@ -1,10 +1,9 @@
 {
   "name": "Guild Folders",
-  "version": "0.0.1",
+  "version": "0.1.0",
   "description": "Allows you to organize your servers with folders and more",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "dependencies": [
     "pc-classNameNormalizer"
   ]
