+++ src/Powercord/plugins/pc-heygirl/manifest.json
@@ -4,7 +4,6 @@
   "description": "heygirl",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "dependencies": [
     "pc-commands"
   ]
