+++ src/Powercord/plugins/pc-keybindManager/manifest.json
@@ -4,7 +4,6 @@
   "description": "Powercord's keybind API",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "dependencies": [
     "pc-settings"
   ]
