+++ src/Powercord/plugins/pc-pluginManager/index.js
@@ -1,10 +1,10 @@
-const Plugin = require('powercord/Plugin');
+const { Plugin } = require('powercord/entities');
 const { resolve } = require('path');
 
 const Settings = require('./components/Settings.jsx');
 
 module.exports = class PluginManager extends Plugin {
-  async start () {
+  async pluginDidLoad () {
     this.loadCSS(resolve(__dirname, 'scss', 'style.scss'));
 
     powercord
@@ -13,7 +13,7 @@ module.exports = class PluginManager extends Plugin {
       .register('pc-pluginManager', 'Plugins', Settings);
   }
 
-  unload () {
+  pluginWillUnload () {
     this.unloadCSS();
     powercord
       .pluginManager
