+++ src/Powercord/plugins/pc-pluginManager/manifest.json
@@ -4,7 +4,6 @@
   "description": "Powercord's plugin manager",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "dependencies": [
     "pc-settings"
   ]
