+++ src/Powercord/plugins/pc-reactdevtools/manifest.json
@@ -4,6 +4,5 @@
   "description": "Adds React Devtools in Discord devtools",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "appMode": "both"
 }
