+++ src/Powercord/plugins/pc-settings/manifest.json
@@ -4,7 +4,6 @@
   "description": "Powercord's settings API",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "appMode": "both",
   "dependencies": [
     "pc-classNameNormalizer"
