+++ src/Powercord/plugins/pc-spotify/manifest.json
@@ -4,7 +4,6 @@
   "description": "Better Spotify integration in Discord",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "dependencies": [
     "pc-commands"
   ]
