+++ src/Powercord/plugins/pc-styleManager/manifest.json
@@ -4,7 +4,6 @@
   "description": "Powercord's style manager",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "appMode": "both",
   "dependencies": [
     "pc-classNameNormalizer"
