+++ src/Powercord/plugins/pc-translate/manifest.json
@@ -3,6 +3,5 @@
   "version": "1.0.0",
   "description": "Quickly translate messages to another lang",
   "author": "Powercord Team",
-  "license": "MIT",
-  "repo": null
+  "license": "MIT"
 }
