+++ src/Powercord/plugins/pc-updater/manifest.json
@@ -4,7 +4,6 @@
   "description": "Ensure Powercord is always up to date",
   "author": "Powercord Team",
   "license": "MIT",
-  "repo": null,
   "dependencies": [
     "pc-settings"
   ]
