+++ src/fake_node_modules/powercord/entities/Plugin.js
@@ -1,7 +1,7 @@
 const Events = require('events');
 const { sleep } = require('powercord/util');
 
-const SettingsManager = require('../../Powercord/settingsManager');
+const SettingsManager = require('../../../Powercord/settingsManager');
 
 module.exports = class Plugin extends Events {
   constructor () {
@@ -13,17 +13,16 @@ module.exports = class Plugin extends Events {
 
   async _start () {
     try {
-      while (!this.manifest.dependencies.every(pluginName => (
-        powercord.pluginManager.get(pluginName).ready
-      ))) {
+      while (!this.manifest.dependencies.every(pluginName => powercord.pluginManager.get(pluginName).ready)) {
         await sleep(1);
       }
 
-      if (typeof this.start === 'function') {
-        await this.start();
+      if (typeof this.pluginDidLoad === 'function') {
+        await this.pluginDidLoad();
       }
+
       this.ready = true;
-      this.log('Plugin started');
+      this.log('Plugin loaded');
     } catch (e) {
       this.error('An error occurred during initialization!', e);
     }
@@ -31,8 +30,8 @@ module.exports = class Plugin extends Events {
 
   async _unload () {
     try {
-      if (typeof this.unload === 'function') {
-        await this.unload();
+      if (typeof this.pluginWillUnload === 'function') {
+        await this.pluginWillUnload();
       }
       this.ready = false;
       this.log('Plugin unloaded');
